Overhaul codebase
As pointed out by a reviewer, the timestepping scheme previously used
is only first order, so we replace the old ins::solver with IB objects
tacked on hackery with a proper solver, ib::solver, which implements the
2nd-order scheme of Peskin (2002). Test files have been updated to use
the new solver and old tests have been removed.

With this update comes a major rearrangement of files in which most
files related to dense solvers, sparse solvers, and preconditioners now
live in the include/solvers directory.

In preparation for release, debugging messages have been hidden behind
the `debug` log level. Printing of debug-level messages can be toggled
by compiling with/out `-DNDEBUG`. Comments have been sprinkled about to
help explain some of the more perplexing parts of the codebase.

Finally, a small bug in the 1D tension code has been corrected.